requirements.txt
pyperclip
plyer
pystray
Pillow